design system table component refactor
