+++ src/applications/debt-letters/components/DebtLettersDownload.jsx
@@ -4,78 +4,12 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import Breadcrumbs from '@department-of-veterans-affairs/component-library/Breadcrumbs';
 import { setPageFocus } from '../utils/page';
-import { MobileTableView } from './MobileTableView';
-import { DebtLettersTable } from './DebtLettersTable';
+import DebtLettersTable from './DebtLettersTable';
 import scrollToTop from 'platform/utilities/ui/scrollToTop';
 import Telephone, {
   CONTACTS,
   PATTERNS,
 } from '@department-of-veterans-affairs/component-library/Telephone';
-
-const ErrorAlert = () => (
-  <div
-    className="usa-alert usa-alert-error vads-u-margin-top--0 vads-u-padding--3"
-    role="alert"
-  >
-    <div className="usa-alert-body">
-      <h3 className="usa-alert-heading">
-        Your debt letters are currently unavailable.
-      </h3>
-      <p className="vads-u-font-family--sans">
-        You can’t download your debt letters because something went wrong on our
-        end.
-      </p>
-      <h4>What you can do</h4>
-      <p className="vads-u-font-family--sans vads-u-margin-y--0">
-        You can check back later or call the Debt Management Center at
-        <Telephone className="vads-u-margin-x--0p5" contact="8008270648" /> to
-        find out more information about how to resolve your debt.
-      </p>
-    </div>
-  </div>
-);
-
-const DependentDebt = () => (
-  <div
-    className="usa-alert usa-alert-error vads-u-margin-top--0 vads-u-padding--3"
-    role="alert"
-  >
-    <div className="usa-alert-body">
-      <h3 className="usa-alert-heading">
-        Your debt letters are currently unavailable.
-      </h3>
-      <p className="vads-u-font-family--sans">
-        You can’t download your debt letters because something went wrong on our
-        end.
-      </p>
-      <h4>What you can do</h4>
-      <p className="vads-u-font-family--sans vads-u-margin-y--0">
-        If you need to access debt letters that were mailed to you, call the
-        Debt Management Center at <Telephone contact="8008270648" />.
-      </p>
-    </div>
-  </div>
-);
-
-const NoDebtLinks = () => (
-  <div className="vads-u-background-color--gray-lightest vads-u-padding--3 vads-u-margin-bottom--2">
-    <div className="usa-alert-body">
-      <h3 className="usa-alert-heading">You don’t have any VA debt letters</h3>
-      <p className="vads-u-font-family--sans">
-        Our records show you don’t have any debt letters related to VA benefits.
-        If you think this is an error, please contact the Debt Management Center
-        at <Telephone contact="8008270648" />.
-      </p>
-      <p className="vads-u-font-family--sans vads-u-margin-y--0">
-        If you have VA health care copay debt, go to our
-        <a className="vads-u-margin-x--0p5" href="/health-care/pay-copay-bill/">
-          Pay your VA copay bill
-        </a>
-        page to learn about your payment options.
-      </p>
-    </div>
-  </div>
-);
 
 const DownloadLettersAlert = () => (
   <va-alert
@@ -111,26 +45,6 @@ const DownloadLettersAlert = () => (
     </p>
   </va-alert>
 );
-
-const DebtLettersView = ({
-  debtLinks,
-  isError,
-  isVBMSError,
-  hasDependentDebts,
-}) => {
-  const hasDebtLinks = !!debtLinks.length;
-
-  if (isError || isVBMSError) return <ErrorAlert />;
-  if (hasDependentDebts) return <DependentDebt />;
-  if (!hasDebtLinks) return <NoDebtLinks />;
-
-  return (
-    <>
-      <DebtLettersTable debtLinks={debtLinks} />
-      <MobileTableView debtLinks={debtLinks} />
-    </>
-  );
-};
 
 const DebtLettersDownload = ({
   debtLinks,
@@ -170,7 +84,7 @@ const DebtLettersDownload = ({
         <DownloadLettersAlert />
 
         <h2>Your debt letters</h2>
-        <DebtLettersView
+        <DebtLettersTable
           debtLinks={debtLinks}
           isError={isError}
           isVBMSError={isVBMSError}
